requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.0.0=pypi_0
addict=2.4.0=pypi_0
ca-certificates=2023.12.12=h06a4308_0
cachetools=5.3.2=pypi_0
certifi=2022.12.7=py37h06a4308_0
charset-normalizer=3.3.2=pypi_0
cycler=0.11.0=pypi_0
easydict=1.13=pypi_0
einops=0.6.1=pypi_0
facexlib=0.3.0=pypi_0
filelock=3.12.2=pypi_0
filterpy=1.4.5=pypi_0
fonttools=4.38.0=pypi_0
fsspec=2023.1.0=pypi_0
ftfy=6.1.1=pypi_0
future=0.18.3=pypi_0
fvcore=0.1.5.post20221221=pypi_0
google-auth=2.25.2=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.60.0=pypi_0
huggingface-hub=0.16.4=pypi_0
idna=3.6=pypi_0
imageio=2.31.2=pypi_0
imgaug=0.4.0=pypi_0
importlib-metadata=6.7.0=pypi_0
iopath=0.1.10=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=2.4=pypi_0
kiwisolver=1.4.5=pypi_0
kornia=0.6.12=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
llvmlite=0.39.1=pypi_0
lmdb=1.4.1=pypi_0
lpips=0.1.4=pypi_0
markdown=3.4.4=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.5.3=pypi_0
natsort=8.4.0=pypi_0
ncurses=6.4=h6a678d5_0
networkx=2.6.3=pypi_0
numba=0.56.4=pypi_0
numpy=1.21.5=pypi_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
oauthlib=3.2.2=pypi_0
openai-clip=1.0.1=pypi_0
opencv-python=4.6.0.66=pypi_0
openssl=1.1.1w=h7f8727e_0
packaging=23.2=pypi_0
pandas=1.3.5=pypi_0
pillow=9.2.0=pypi_0
pip=22.3.1=py37h06a4308_0
platformdirs=4.0.0=pypi_0
portalocker=2.7.0=pypi_0
protobuf=3.20.3=pypi_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pyiqa=0.1.10=pypi_0
pyparsing=3.1.2=pypi_0
python=3.7.16=h7a1cb2a_0
python-dateutil=2.8.2=pypi_0
pytorch-msssim=1.0.0=pypi_0
pytz=2023.3.post1=pypi_0
pywavelets=1.3.0=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
regex=2022.8.17=pypi_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.1=pypi_0
scikit-image=0.19.3=pypi_0
scipy=1.7.3=pypi_0
setuptools=63.4.1=pypi_0
shapely=2.0.4=pypi_0
six=1.16.0=pypi_0
sqlite=3.41.2=h5eee18b_0
tabulate=0.9.0=pypi_0
tb-nightly=2.12.0a20230113=pypi_0
tensorboard=2.11.2=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorboardx=2.6.2.2=pypi_0
termcolor=2.3.0=pypi_0
tifffile=2021.11.2=pypi_0
timm=0.9.12=pypi_0
tk=8.6.12=h1ccaba5_0
tomli=2.0.1=pypi_0
torch=1.13.1=pypi_0
torchvision=0.14.1=pypi_0
tornado=6.2=pypi_0
tqdm=4.64.0=pypi_0
typing-extensions=4.7.1=pypi_0
urllib3=2.0.7=pypi_0
visdom=0.2.4=pypi_0
wcwidth=0.2.12=pypi_0
websocket-client=1.6.1=pypi_0
werkzeug=2.2.3=pypi_0
wheel=0.38.4=py37h06a4308_0
xz=5.4.5=h5eee18b_0
yacs=0.1.8=pypi_0
yapf=0.40.2=pypi_0
zipp=3.15.0=pypi_0
zlib=1.2.13=h5eee18b_0
